beam_batch_pipeline/requirements.txt
tweepy
kafka-python
apache-beam
nltk
